week4/community-contributions/docstring_unity_test_tool/pyproject.toml
[project]
name = "docstring-unity-test-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "astor>=0.8.1",
    "gradio>=5.49.1",
    "openai>=2.6.1",
    "openai-agents>=0.4.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "typer>=0.20.0",
]
